added handle to rect out of image case

+++ cvat-ui/src/utils/opencv-wrapper/tracker-mil.ts
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
+import { convertPointInBounds } from 'utils/math';
 import { TrackerModel, TrackingResult } from './opencv-interfaces';
 
 export type TrackerMIL = TrackerModel;
@@ -21,20 +22,30 @@ export default class TrackerMILImplementation implements TrackerMIL {
 
     public init(src: ImageData, points: number[]): void {
         if (points.length !== 4) {
-            throw Error(`TrackerMIL must be initialized with rect, but got ${points.length % 2} points.`);
+            throw Error(`TrackerMIL must be initialized with rectangle, but got ${points.length % 2} points.`);
         }
         this.prevImageData = src;
-        const [x, y, width, height] = [points[0], points[1], points[2] - points[0], points[3] - points[1]];
-        const matImage = this.cv.matFromImageData(src);
-        const rect = new this.cv.Rect(x, y, width, height);
-        this.trackerMIL.init(matImage, rect);
-        matImage.delete();
+
+        // cut bounding box if its out of image bounds
+        const { x: x1, y: y1 } = convertPointInBounds(points[0], points[1], src.width, src.height);
+        const { x: x2, y: y2 } = convertPointInBounds(points[2], points[3], src.width, src.height);
+
+        const [width, height] = [x2 - x1, y2 - y1];
+        if (width === 0 || height === 0) {
+            throw Error('TrackerMIL got rectangle out of image bounds');
+        }
+
+        let matImage = null;
+        try {
+            matImage = this.cv.matFromImageData(src);
+            const rect = new this.cv.Rect(x1, y1, width, height);
+            this.trackerMIL.init(matImage, rect);
+        } finally {
+            if (matImage) matImage.delete();
+        }
     }
 
     public reinit(points: number[]): void {
-        if (points.length !== 4) {
-            throw Error(`TrackerMIL must be initialized with rect, but got ${points.length % 2} points.`);
-        }
         if (!this.prevImageData) {
             throw Error('TrackerMIL needs to be initialized before re-initialization');
         }
@@ -43,9 +54,13 @@ export default class TrackerMILImplementation implements TrackerMIL {
 
     public update(src: ImageData): TrackingResult {
         this.prevImageData = src;
-        const matImage = this.cv.matFromImageData(src);
-        const [updated, rect] = this.trackerMIL.update(matImage);
-        matImage.delete();
-        return { updated, points: [rect.x, rect.y, rect.x + rect.width, rect.y + rect.height] };
+        let matImage = null;
+        try {
+            matImage = this.cv.matFromImageData(src);
+            const [updated, rect] = this.trackerMIL.update(matImage);
+            return { updated, points: [rect.x, rect.y, rect.x + rect.width, rect.y + rect.height] };
+        } finally {
+            if (matImage) matImage.delete();
+        }
     }
 }
